WIP: implementing enrollment

+++ src/main/java/org/example/model/Lecturer.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 public class Lecturer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Integer id;
+    public Long id;
 
     @Column(name = "user_id", nullable = false)
     public Long userId;
@@ -34,9 +34,8 @@ public class Lecturer {
 
     public Lecturer() {}
 
-    // Getters & Setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public Long getUserId() { return userId; }
     public void setUserId(Long userId) { this.userId = userId; }
