+++ src/main/java/org/example/repository/LecturerRepository.java
@@ -3,4 +3,4 @@ package org.example.repository;
 import org.example.model.Lecturer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LecturerRepository extends JpaRepository<Lecturer, Integer> {}
+public interface LecturerRepository extends JpaRepository<Lecturer, Long> {}
