+++ src/main/java/org/example/repository/StudentRepository.java
@@ -3,4 +3,4 @@ package org.example.repository;
 import org.example.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {}
+public interface StudentRepository extends JpaRepository<Student, Long> {}
